Fixed Token Recogntion
I found that the program would experience a segmentation fault on a different computer
while parsing through the tokens. I resolved this by using a list terminator that
should work on any system; instead of NULL, which for some reason takes on its
address, I opted to use "" instead. Since its only character is \0, it is easier
to test for it with strcmp.