Small Tweak
I have modified the program so that after leading spaces have been removed,
tokens or variable names must be the first value on the line. Also, I have
changed the Makefile so that the executable is now called ascript instead
of testfile. I found this change to be somewhat obligatory.