Converted PEMDAS To Stacks
I was advised to use the native stacks in the ATmega328P assembly, which
allows me to push and pop bytes from a natively available stack. The result
is code that is expected to be more efficient, as well as shorter in
terms if the number of instructions (which should conserve memory).