Changed Comparison Branching
During my first test run, I noticed that brgt and brle are not commands
in the Arduino assembly, since brlt and brge can be used to do the same
thing if you reverse the parameters. So, I fixed a couple of the comparisons,
which should fix any issues with this.