Switched Data Scheme To Stack Frame
Knowing that I will eventually make use of stack frames, I have
decided to use the stack to store variables. The new scheme
involves having variables get pushed to the stack at the beginning
of execution. The program can then access these values safely.
In the beginning, I was pushing variables and calculations to the
lower end of RAM. However, this change completes the flip in the
position of everything.