Function Skip Jump
I have added jmp statements on the outside of functions that
cause a function to automatically be skipped. I could use this
in several ways. First, I could use this to ensure that functions
have to be called outside of the scope of any function (for instance,
in the global scope through a function; think Python). Another
route could be to make an exception with a main() function,
which would not have the jump. This would cause main() to be
automatically invoked upon execution.