Fixed Parameter Artifacts
I found an issue in the parameters where strings outside of the
string's memory were being accessed. The issue is resolved.
I have also discovered the limitations of the branch statements,
which as it turns out can only jump a short distance. My intention
is to replace this with a jump over something along the lines of
a branch over a jump statement. I'll have to consider the options.