Fixed Stack Reset Issue
I realized through testing that the stack frames were not
resetting properly due to the use of subtraction instead of
addition. I must have been tricked by the wording. The issue
is now resolved, and thus function calls are finished. I also
noticed that the summation example was incorrect. It was
multiplying instead of adding. This is no longer the case.